Neatened up DistanceMatrix Functionality, removing functions we are no longer using and the output to the correct format.

+++ src/CMap_test.cpp
@@ -52,7 +52,7 @@ int CMap_test()
 	return 0;
 }
 
-void PrintMatrix(std::vector<std::vector<int>> inputMap)
+void PrintMatrix(std::vector<std::vector<double>> inputMap)
 {
 	DEBUG_METHOD();
